feat:P4ADEV-1545-FdRIngestionActity-implementation-retrieval-via-query-file-validation - moved package
