+++ src/main/java/it/gov/pagopa/payhub/activities/activity/paymentsreporting/ReportingFlowIngestionActivityImpl.java
@@ -1,10 +1,10 @@
 package it.gov.pagopa.payhub.activities.activity.paymentsreporting;
 
-import it.gov.pagopa.payhub.activities.service.IngestionFileHandlerService;
-import it.gov.pagopa.payhub.activities.service.IngestionFileValidatorService;
-import it.gov.pagopa.payhub.activities.service.IngestionFlowRetrieverService;
+import it.gov.pagopa.payhub.activities.service.ingestionflow.IngestionFileHandlerService;
+import it.gov.pagopa.payhub.activities.service.ingestionflow.IngestionFileValidatorService;
+import it.gov.pagopa.payhub.activities.service.ingestionflow.IngestionFlowRetrieverService;
 import it.gov.pagopa.payhub.activities.dto.reportingflow.ReportingFlowIngestionActivityResult;
-import it.gov.pagopa.payhub.activities.dto.reportingflow.IngestionFlowDTO;
+import it.gov.pagopa.payhub.activities.dto.ingestionflow.IngestionFlowDTO;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
