+++ src/main/java/it/gov/pagopa/payhub/activities/activity/utility/SendEmailIngestionFlowActivity.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payhub.activities.activity.paymentsreporting;
+package it.gov.pagopa.payhub.activities.activity.utility;
 
 /**
  * Interface for SendEmailIngestionFlowActivity.
