+++ src/main/java/it/gov/pagopa/payhub/activities/service/ingestionflow/IngestionFileValidatorService.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payhub.activities.service;
+package it.gov.pagopa.payhub.activities.service.ingestionflow;
 
 import it.gov.pagopa.payhub.activities.exception.InvalidIngestionFileException;
 import lombok.extern.slf4j.Slf4j;
