+++ src/test/java/it/gov/pagopa/payhub/activities/service/IngestionFileValidatorServiceTest.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.payhub.activities.service;
 
 import it.gov.pagopa.payhub.activities.exception.InvalidIngestionFileException;
-import it.gov.pagopa.payhub.activities.service.IngestionFileValidatorService;
+import it.gov.pagopa.payhub.activities.service.ingestionflow.IngestionFileValidatorService;
 import org.apache.commons.codec.binary.Hex;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
